volequ.sh: not just ogg
Update this script (volequ.sh) so that it is not ogg-specific; make it
flexible enough to be able to process sound files of any type (e.g. mp3).

All that is really required is to not assume '.ogg' as the input filename
extension and simply find the extension by examining the input file.

+++ bin/volequ.sh
@@ -12,20 +12,26 @@ fi
 FILECNT=$#
 
 for (( I=1 ; I <= $FILECNT ; I=`expr $I + 1` )); do
-	TMPOUT=`mktemp out.XXXXXXXXXX`
-	if [ $? -ne 0 ]; then
-		echo "can't create tmp file"
-		exit 1
-	fi
-	rm -f $TMPOUT
-
 	echo -n "working on file: \"$1\": "
-	TMP=`sox "$1" -n stat -v 2>&1`
-	echo "$TMP "
 
-	if [ "$TMP" != "1.000" ]; then
-		sox -SV "$1" "${TMPOUT}.ogg" vol $TMP
-		mv "${TMPOUT}.ogg" "$1"
+	FILEEXT=$(echo "$1" | rev | cut -d'.' -f1 | rev)
+	if [ x"$FILEEXT" = x ]; then
+		echo "input file ($1) has no extension"
+	else
+		TMPOUT=$(mktemp out.XXXXXXXXXX)
+		if [ $? -ne 0 ]; then
+			echo "can't create tmp file"
+			exit 1
+		fi
+		rm -f $TMPOUT
+
+		TMP=`sox "$1" -n stat -v 2>&1`
+		echo "$TMP "
+
+		if [ "$TMP" != "1.000" ]; then
+			sox -SV "$1" "${TMPOUT}.${FILEEXT}" vol $TMP
+			mv "${TMPOUT}.${FILEEXT}" "$1"
+		fi
 	fi
 
 	# next file
